backend/requirements.txt
SQLModel ~= 0.0.8
sqlalchemy==1.4.35 # https://github.com/tiangolo/sqlmodel/issues/383#issuecomment-1193373160
FastAPI[all] ~= 0.95.1
alembic ~= 1.10.4
pydantic[email] ~= 1.10.7
python-jose[cryptography] ~= 3.3.0 
pyjwt[crypto] ~= 2.6.0
passlib[bcrypt] ~= 1.7.4
python-dotenv ~= 1.0.0
uvicorn[standard] ~= 0.20.0
gunicorn ~= 20.1.0
psycopg2-binary ~= 2.9.6
